test(daemon): move the zcode api-key instance test into its own contract file
